% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmc.R
\name{tmc}
\alias{tmc}
\title{Compute Transitive Maximal Correlation Matrix}
\usage{
tmc(cor_mat)
}
\arguments{
\item{cor_mat}{A correlation matrix.}
}
\value{
A list containing:
\item{tmc_matrix}{The transitive maximal correlation matrix.}
\item{attractors}{A list of attractor pairs.}
\item{attractor_basins}{A list of attractor basin compositions.}
}
\description{
This function takes as input a correlation matrix and returns a list containing the corresponding transitive maximal correlation matrix, a list of attractor pairs, and a list of attractor basin compositions.
}
